requirements.txt
# Extra wheel repository that contains the CUDA 12.4 / PyTorch 2.4 wheels
--find-links https://data.pyg.org/whl/torch-2.4.0+cu124.html

absl-py==2.1.0
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
ase==3.23.0
attrs==24.2.0
beartype==0.20.2
bitsandbytes==0.43.3
blosc2==3.2.1
braceexpand==0.1.7
charset-normalizer==3.4.0
chgnet==0.3.1
click==8.1.7
cloudpickle==3.1.0
ConfigArgParse==1.7
contourpy==1.3.1
cut-cross-entropy==25.1.1
cycler==0.12.1
Cython==3.0.12
datasets==2.21.0
dgl==2.4.0+cu121
diffusers==0.32.2
dill==0.3.8
dnspython==2.7.0
docker-pycreds==0.4.0
docstring_parser==0.16
e3nn==0.5.6
einops==0.8.1
fairchem-core==1.10.0
filelock==3.16.1
flash_attn==2.7.4.post1
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.70.0
h5py==3.13.0
hf_transfer==0.1.9
huggingface-hub==0.30.1
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.5.0
jaxtyping==0.3.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
latexcodec==3.0.0
lightning==2.4.0
lightning-utilities==0.11.9
llvmlite==0.44.0
lmdb==1.6.2
lxml==5.3.0
mace-torch==0.3.12
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matgl==1.1.3
matminer==0.9.3
matplotlib==3.9.3
matscipy==1.1.1
mdurl==0.1.2
monty==2024.10.21
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy_extensions==1.1.0
ndindex==1.9.2
networkx==3.4.2
numba==0.61.0
numexpr==2.10.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opt-einsum-fx==0.1.4
opt_einsum==3.4.0
orjson==3.10.15
p_tqdm==1.4.2
packaging==24.2
palettable==3.3.3
pandas==2.2.3
pathos==0.3.3
peft==0.12.0
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
pox==0.3.5
ppft==1.7.6.9
prettytable==3.16.0
propcache==0.2.1
protobuf==3.20.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==18.1.0
pybtex==0.24.0
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.19.1
pymatgen==2024.8.9
pymongo==4.10.1
pyocse==0.1.2
pyparsing==3.2.0
pyre-extensions==0.0.32
python-dateutil==2.9.0.post0
python-hostlist==2.2.1
pytorch-lightning==2.4.0
pytz==2024.2
pyxtal==1.0.6
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
sevenn @ git+https://github.com/MDIL-SNU/SevenNet.git@8ce2c9d4e2aa6e5fe19bcf0882cbaab82a6ec0e8
shtab==1.7.1
six==1.17.0
SMACT==3.0.1
smmap==5.0.1
spglib==2.5.0
submitit==1.5.2
sympy==1.13.1
tables==3.10.2
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
timeout-decorator==0.5.0
tokenizers==0.21.0
toml==0.10.2
torch==2.4.0
torch-ema==0.3
torch-geometric==2.6.1
torch_cluster==1.6.3
torch_scatter==2.1.2
torch_sparse==0.6.18
torchaudio==2.4.0
torchdata==0.7.1
torchmetrics==1.6.0
torchtnt==0.2.4
torchvision==0.19.0
tqdm==4.67.1
transformers==4.48.0
triton==3.0.0
trl==0.13.0
typeguard==4.4.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tyro==0.9.17
tzdata==2024.2
uncertainties==3.2.2
unsloth==2025.3.18
unsloth_zoo==2025.3.16
urllib3==2.2.3
vasprun-xml==1.0.4
vesin==0.3.2
vesin-torch==0.3.2
wadler_lindig==0.1.5
wandb==0.17.7
wcwidth==0.2.13
Werkzeug==3.1.3
xformers==0.0.27.post2
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
